新建页目录

+++ kernel/include/kernel/mm.h
@@ -54,8 +54,8 @@ void vmm_cleanup_low_identity_mapping();
 
 uintptr_t vmm_alloc_pages(uint32_t size, uint32_t flag);
 
-static inline uint32_t vmm_get_cr3() {
-    uint32_t cr3;
+static inline uintptr_t vmm_get_cr3() {
+    uintptr_t cr3;
     asm volatile("mov %%cr3, %0" : "=r"(cr3));
     return cr3;
 }
